+++ __tests__/ApplicationTest.js
@@ -1,5 +1,5 @@
 const MissionUtils = require("@woowacourse/mission-utils");
-const App = require("../src/classes/App");
+const App = require("../src/App");
 const BridgeMaker = require("../src/BridgeMaker");
 
 const mockQuestions = (answers) => {
