+++ docs/README.md
@@ -5,6 +5,21 @@
 - 먼저 기능이 작동하게 구현 -> 리팩토링 & TDD
   <br/>
 
+## 리팩토링 & 제출 전 점검 요소
+- [x] MVC 패턴대로 분류하기
+- [ ] 변수 정리
+- [ ] type 주석 추가
+- [ ] 예외는 throw 문을 사용해 발생, "[ERROR] ..." + 바로 종료가 아니라 그 부분부터 입력을 다시 받는다
+- [ ] ApplicationTest PASS 여부
+- [ ] 요구사항에서 명시하지 않는 한 파일, 패키지 이름 수정, 이동 금지
+- [ ] indent 2까지
+- [ ] 함수가 하나의 일만 하도록, 10라인 미만
+- [ ] else 지양
+- [ ] 파라미터는 최대 3개까지
+- [ ] InputView 에서만 입력받기- readLind()
+- [ ] BridgeGame 클래스에서 InputView, OutputView 사용 금지
+
+##
 ## 🌉 기능 구현 목록
 
 ###
