+++ src/Controller/Controller.js
@@ -1,0 +1,3 @@
+class Controller {}
+
+module.exports = { Controller };
