+++ src/Model/BridgeGame.js
@@ -1,4 +1,4 @@
-const Check = require("../utils/Check");
+const Check = require("../Utils/Check");
 
 /**
  * 다리 건너기 게임을 관리하는 클래스 -> InputView, OutputView 사용 불가
