--- src/classes/Test.js
@@ -1,4 +1,0 @@
-const App = require("./App");
-
-const app = new App();
-app.test();
